refactor: Move infrastructure service routes to a new component
Start moving infrastructure service related code to a dedicated
component by moving the routes to a new component.
