+++ components/infrastructure-services/api-model/src/commonMain/kotlin/UpdateInfrastructureService.kt
@@ -1,0 +1,37 @@
+/*
+ * Copyright (C) 2023 The ORT Server Authors (See <https://github.com/eclipse-apoapsis/ort-server/blob/main/NOTICE>)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ * License-Filename: LICENSE
+ */
+
+package org.eclipse.apoapsis.ortserver.components.infrastructureservices
+
+import kotlinx.serialization.Serializable
+
+import org.eclipse.apoapsis.ortserver.shared.apimodel.CredentialsType
+import org.eclipse.apoapsis.ortserver.shared.apimodel.OptionalValue
+
+/**
+ * Request object for the update infrastructure service endpoint.
+ */
+@Serializable
+data class UpdateInfrastructureService(
+    val url: OptionalValue<String> = OptionalValue.Absent,
+    val description: OptionalValue<String?> = OptionalValue.Absent,
+    val usernameSecretRef: OptionalValue<String> = OptionalValue.Absent,
+    val passwordSecretRef: OptionalValue<String> = OptionalValue.Absent,
+    val credentialsTypes: OptionalValue<Set<CredentialsType>> = OptionalValue.Absent
+)
