+++ components/infrastructure-services/backend/src/routes/kotlin/Validation.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2024 The ORT Server Authors (See <https://github.com/eclipse-apoapsis/ort-server/blob/main/NOTICE>)
+ * Copyright (C) 2023 The ORT Server Authors (See <https://github.com/eclipse-apoapsis/ort-server/blob/main/NOTICE>)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,23 +17,14 @@
  * License-Filename: LICENSE
  */
 
-plugins {
-    // Apply precompiled plugins.
-    id("ort-server-kotlin-multiplatform-conventions")
-    id("ort-server-publication-conventions")
-}
+package org.eclipse.apoapsis.ortserver.components.infrastructureservices
 
-group = "org.eclipse.apoapsis.ortserver.api.v1"
+import io.ktor.server.plugins.requestvalidation.RequestValidationConfig
 
-kotlin {
-    sourceSets {
-        commonMain {
-            dependencies {
-                api(projects.api.v1.apiV1Model)
-                api(projects.model)
+import org.eclipse.apoapsis.ortserver.shared.ktorutils.mapValidationResult
 
-                implementation(projects.shared.apiMappings)
-            }
-        }
+fun RequestValidationConfig.infrastructureServicesValidations() {
+    validate<CreateInfrastructureService> { create ->
+        mapValidationResult(CreateInfrastructureService.validate(create))
     }
 }
