+++ core/src/main/kotlin/plugins/OpenApi.kt
@@ -46,9 +46,9 @@ import io.ktor.server.routing.routing
 import kotlinx.datetime.Instant
 import kotlinx.serialization.json.Json
 
-import org.eclipse.apoapsis.ortserver.api.v1.model.CredentialsType
 import org.eclipse.apoapsis.ortserver.api.v1.model.RepositoryType
 import org.eclipse.apoapsis.ortserver.components.authorization.SecurityConfigurations
+import org.eclipse.apoapsis.ortserver.shared.apimodel.CredentialsType
 import org.eclipse.apoapsis.ortserver.shared.apimodel.OptionalValue
 import org.eclipse.apoapsis.ortserver.utils.system.ORT_SERVER_VERSION
