+++ core/src/main/kotlin/plugins/Routing.kt
@@ -26,6 +26,7 @@ import io.ktor.server.routing.routing
 
 import org.eclipse.apoapsis.ortserver.components.adminconfig.adminConfigRoutes
 import org.eclipse.apoapsis.ortserver.components.authorization.SecurityConfigurations
+import org.eclipse.apoapsis.ortserver.components.infrastructureservices.infrastructureServicesRoutes
 import org.eclipse.apoapsis.ortserver.components.pluginmanager.pluginManagerRoutes
 import org.eclipse.apoapsis.ortserver.components.secrets.secretsRoutes
 import org.eclipse.apoapsis.ortserver.compositions.secretsroutes.secretsCompositionRoutes
@@ -50,6 +51,7 @@ fun Application.configureRouting() {
             authenticate(SecurityConfigurations.TOKEN) {
                 admin()
                 adminConfigRoutes(get())
+                infrastructureServicesRoutes(get())
                 organizations()
                 pluginManagerRoutes(get(), get(), get())
                 products()
