+++ settings.gradle.kts
@@ -32,6 +32,8 @@ include(":components:admin-config:api-model")
 include(":components:admin-config:backend")
 include(":components:authorization:api-model")
 include(":components:authorization:backend")
+include(":components:infrastructure-services:api-model")
+include(":components:infrastructure-services:backend")
 include(":components:plugin-manager:api-model")
 include(":components:plugin-manager:backend")
 include(":components:secrets:api-model")
