+++ shared/api-mappings/src/commonMain/kotlin/InfrastructureServiceMappings.kt
@@ -1,0 +1,64 @@
+/*
+ * Copyright (C) 2023 The ORT Server Authors (See <https://github.com/eclipse-apoapsis/ort-server/blob/main/NOTICE>)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ * License-Filename: LICENSE
+ */
+
+package org.eclipse.apoapsis.ortserver.shared.apimappings
+
+import org.eclipse.apoapsis.ortserver.model.CredentialsType
+import org.eclipse.apoapsis.ortserver.model.InfrastructureService
+import org.eclipse.apoapsis.ortserver.model.InfrastructureServiceDeclaration
+import org.eclipse.apoapsis.ortserver.shared.apimodel.CredentialsType as ApiCredentialsType
+import org.eclipse.apoapsis.ortserver.shared.apimodel.InfrastructureService as ApiInfrastructureService
+
+fun CredentialsType.mapToApi() = ApiCredentialsType.valueOf(name)
+
+fun Set<CredentialsType>.mapToApi(): Set<ApiCredentialsType> = mapTo(mutableSetOf()) { it.mapToApi() }
+
+fun ApiCredentialsType.mapToModel() = CredentialsType.valueOf(name)
+
+fun Set<ApiCredentialsType>.mapToModel(): Set<CredentialsType> = mapTo(mutableSetOf()) { it.mapToModel() }
+
+fun InfrastructureService.mapToApi() =
+    ApiInfrastructureService(
+        name = name,
+        url = url,
+        description = description,
+        usernameSecretRef = usernameSecret.name,
+        passwordSecretRef = passwordSecret.name,
+        credentialsTypes = credentialsTypes.mapToApi()
+    )
+
+fun ApiInfrastructureService.mapToModel() =
+    InfrastructureServiceDeclaration(
+        name = name,
+        url = url,
+        description = description,
+        usernameSecret = usernameSecretRef,
+        passwordSecret = passwordSecretRef,
+        credentialsTypes = credentialsTypes.mapToModel()
+    )
+
+fun InfrastructureServiceDeclaration.mapToApi() =
+    ApiInfrastructureService(
+        name = name,
+        url = url,
+        description = description,
+        usernameSecretRef = usernameSecret,
+        passwordSecretRef = passwordSecret,
+        credentialsTypes = credentialsTypes.mapToApi()
+    )
